refactor alsa thread to use SoundCallback trait
